third_party/WebKit/LayoutTests/bluetooth/characteristic/stopNotifications/gen-gatt-op-disconnect-called-before.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'disconnect() called before stopNotifications. ' +
    'Reject with NetworkError.';
const value = new Uint8Array([1]);
const expected = new DOMException('GATT Server is disconnected. Cannot ' +
    'perform GATT operations. (Re)connect first with `device.gatt.connect`.',
    'NetworkError')
let device, characteristic, fake_peripheral;

bluetooth_test(() => getMeasurementIntervalCharacteristic()
    .then(_ => ({device, characteristic, fake_peripheral} = _))
    .then(() => {
      device.gatt.disconnect();
      return assert_promise_rejects_with_message(
        characteristic.stopNotifications(),
        expected);
    }),
    test_desc);

</script>
